requirements.txt
stellargraph
pandas
numpy
scikit-learn
scipy
biopython
matplotlib
tensorflow
